Added Lab (WIP) Moved to doubles to avoid casting to float all the time Moving logic to types Tests done using percentages, not absolute values

+++ src/main/java/kofa/maths/Curve.java
@@ -1,24 +1,22 @@
 package kofa.maths;
 
-import org.checkerframework.checker.units.qual.C;
-
 import static com.google.common.base.Preconditions.checkArgument;
 
 // https://discuss.pixls.us/t/tone-curve-math-question/28978/6
 public class Curve {
-    private final float gradient;
-    private final float shoulderStart;
-    private final float mappedValueOfShoulderStart;
-    private final float mappedValueOfShoulderStartMinusOne;
+    private final double gradient;
+    private final double shoulderStart;
+    private final double mappedValueOfShoulderStart;
+    private final double mappedValueOfShoulderStartMinusOne;
 
-    public static Curve linearUntil(float shoulderStart) {
+    public static Curve linearUntil(double shoulderStart) {
         return new Curve(1, shoulderStart);
     }
 
-    public Curve(float gradient, float shoulderStart) {
-        checkArgument(gradient > 0f, "gradient must be > 0");
-        checkArgument(shoulderStart >= 0f, "shoulderStart must be >= 0");
-        float shoulderStartUpperBound = 1 / gradient;
+    public Curve(double gradient, double shoulderStart) {
+        checkArgument(gradient > 0, "gradient must be > 0");
+        checkArgument(shoulderStart >= 0, "shoulderStart must be >= 0");
+        double shoulderStartUpperBound = 1 / gradient;
         checkArgument(shoulderStart < shoulderStartUpperBound, "shoulderStart must be < 1 / gradient = %s", shoulderStartUpperBound);
         this.gradient = gradient;
         this.shoulderStart = shoulderStart;
@@ -26,11 +24,11 @@ public class Curve {
         mappedValueOfShoulderStartMinusOne = mappedValueOfShoulderStart - 1;
     }
 
-    public float mappedValueOf(float x) {
+    public double mappedValueOf(double x) {
         if (x < shoulderStart) {
             return gradient * x;
         }
 
-        return (float) (1 + mappedValueOfShoulderStartMinusOne * Math.exp((gradient * x - mappedValueOfShoulderStart) / mappedValueOfShoulderStartMinusOne));
+        return (1 + mappedValueOfShoulderStartMinusOne * Math.exp((gradient * x - mappedValueOfShoulderStart) / mappedValueOfShoulderStartMinusOne));
     }
 }
